moved DappToWalletInteraction to separate file in messages
