+++ app/src/main/java/com/babylon/wallet/android/domain/usecases/deeplink/ProcessDeepLinkUseCase.kt
@@ -2,7 +2,7 @@ package com.babylon.wallet.android.domain.usecases.deeplink
 
 import com.babylon.wallet.android.data.dapp.IncomingRequestRepository
 import com.babylon.wallet.android.data.dapp.model.toDomainModel
-import com.babylon.wallet.android.domain.model.messages.IncomingMessage
+import com.babylon.wallet.android.domain.model.messages.DappToWalletInteraction
 import com.babylon.wallet.android.domain.model.messages.RemoteEntityID
 import com.radixdlt.sargon.RadixConnectMobile
 import rdx.works.profile.domain.GetProfileUseCase
@@ -38,5 +38,5 @@ class ProcessDeepLinkUseCase @Inject constructor(
 
 sealed interface DeepLinkProcessingResult {
     data object Buffered : DeepLinkProcessingResult
-    data class Processed(val request: IncomingMessage.DappToWalletInteraction) : DeepLinkProcessingResult
+    data class Processed(val request: DappToWalletInteraction) : DeepLinkProcessingResult
 }
