+++ app/src/main/java/com/babylon/wallet/android/presentation/main/MainViewModel.kt
@@ -7,7 +7,7 @@ import com.babylon.wallet.android.data.dapp.IncomingRequestRepository
 import com.babylon.wallet.android.data.dapp.PeerdroidClient
 import com.babylon.wallet.android.data.repository.p2plink.P2PLinksRepository
 import com.babylon.wallet.android.domain.RadixWalletException
-import com.babylon.wallet.android.domain.model.messages.IncomingMessage
+import com.babylon.wallet.android.domain.model.messages.DappToWalletInteraction
 import com.babylon.wallet.android.domain.usecases.AuthorizeSpecifiedPersonaUseCase
 import com.babylon.wallet.android.domain.usecases.VerifyDAppUseCase
 import com.babylon.wallet.android.domain.usecases.deeplink.DeepLinkProcessingResult
@@ -313,7 +313,7 @@ class MainViewModel @Inject constructor(
         }
     }
 
-    private fun verifyIncomingRequest(request: IncomingMessage.DappToWalletInteraction) {
+    private fun verifyIncomingRequest(request: DappToWalletInteraction) {
         verifyingDappRequestJob = viewModelScope.launch {
             verifyDappUseCase(request).onSuccess { verified ->
                 if (verified) {
@@ -413,7 +413,7 @@ class MainViewModel @Inject constructor(
 }
 
 sealed class MainEvent : OneOffEvent {
-    data class IncomingRequestEvent(val request: IncomingMessage.DappToWalletInteraction) : MainEvent()
+    data class IncomingRequestEvent(val request: DappToWalletInteraction) : MainEvent()
 }
 
 data class MainUiState(
