+++ app/src/test/java/com/babylon/wallet/android/data/dapp/DappToWalletInteractionRepositoryTest.kt
@@ -1,7 +1,7 @@
 package com.babylon.wallet.android.data.dapp
 
+import com.babylon.wallet.android.domain.model.messages.DappToWalletInteraction
 import com.babylon.wallet.android.domain.model.messages.WalletAuthorizedRequest
-import com.babylon.wallet.android.domain.model.messages.IncomingMessage.DappToWalletInteraction
 import com.babylon.wallet.android.domain.model.messages.RemoteEntityID
 import com.babylon.wallet.android.utils.AppEvent
 import com.babylon.wallet.android.utils.AppEventBusImpl
