+++ app/src/test/java/com/babylon/wallet/android/presentation/dapp/authorized/personaonetime/PersonaDataOnetimeViewModelTest.kt
@@ -2,6 +2,7 @@ package com.babylon.wallet.android.presentation.dapp.authorized.personaonetime
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
+import com.babylon.wallet.android.domain.model.messages.DappToWalletInteraction
 import com.babylon.wallet.android.domain.model.messages.RequiredPersonaField
 import com.babylon.wallet.android.domain.model.messages.RequiredPersonaFields
 import com.babylon.wallet.android.domain.model.messages.IncomingMessage
@@ -61,9 +62,9 @@ internal class PersonaDataOnetimeViewModelTest : StateViewModelTest<PersonaDataO
             fields = listOf(
                 RequiredPersonaField(
                     PersonaDataField.Kind.Name,
-                    IncomingMessage.DappToWalletInteraction.NumberOfValues(
+                    DappToWalletInteraction.NumberOfValues(
                         1,
-                        IncomingMessage.DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
+                        DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
                     )
                 )
             )
