+++ app/src/test/java/com/babylon/wallet/android/presentation/dapp/authorized/personaongoing/PersonaDataOngoingViewModelTest.kt
@@ -4,6 +4,7 @@ package com.babylon.wallet.android.presentation.dapp.authorized.personaongoing
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
+import com.babylon.wallet.android.domain.model.messages.DappToWalletInteraction
 import com.babylon.wallet.android.domain.model.messages.RequiredPersonaField
 import com.babylon.wallet.android.domain.model.messages.RequiredPersonaFields
 import com.babylon.wallet.android.domain.model.messages.IncomingMessage
@@ -95,9 +96,9 @@ internal class PersonaDataOngoingViewModelTest {
             fields = listOf(
                 RequiredPersonaField(
                     PersonaDataField.Kind.Name,
-                    IncomingMessage.DappToWalletInteraction.NumberOfValues(
+                    DappToWalletInteraction.NumberOfValues(
                         1,
-                        IncomingMessage.DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
+                        DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
                     )
                 )
             )
@@ -124,9 +125,9 @@ internal class PersonaDataOngoingViewModelTest {
             fields = listOf(
                 RequiredPersonaField(
                     PersonaDataField.Kind.PhoneNumber,
-                    IncomingMessage.DappToWalletInteraction.NumberOfValues(
+                    DappToWalletInteraction.NumberOfValues(
                         1,
-                        IncomingMessage.DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
+                        DappToWalletInteraction.NumberOfValues.Quantifier.Exactly
                     )
                 )
             )
